improve tmux script

+++ scripts/tmux.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 
-wget https://raw.githubusercontent.com/dbldots/dotfiles/master/files/_tmux.conf -O ~/.johannes.tmux.conf
-tmux -L johannes new-session -d
-tmux -L johannes send-keys 'tmux source-file ~/.johannes.tmux.conf' C-j
-tmux -L johannes attach
+if ! tmux -L johannes attach ; then
+    wget https://raw.githubusercontent.com/dbldots/dotfiles/master/files/_tmux.conf -O ~/.johannes.tmux.conf
+    tmux -L johannes new-session -d
+    tmux -L johannes send-keys 'tmux source-file ~/.johannes.tmux.conf' C-j
+    tmux -L johannes attach
+fi
